requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
openai>=1.0.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.35.0
Pillow>=10.0.0
bcrypt>=4.0.0
pydantic>=2.0.0

